create a type for mapgen post-processors
