+++ src/overmapbuffer.h
@@ -214,6 +214,7 @@ class overmapbuffer
         const oter_id &ter_existing( const tripoint_abs_omt &p );
         void ter_set( const tripoint_abs_omt &p, const oter_id &id );
         std::optional<mapgen_arguments> *mapgen_args( const tripoint_abs_omt & );
+        pp_generator *get_generator_var( const map_generator &mg );
         std::string *join_used_at( const std::pair<tripoint_abs_omt, cube_direction> & );
         std::vector<oter_id> predecessors( const tripoint_abs_omt & );
         // pick an OMT, scan it from z level 10, and return the first level that is not air
